Code review changes + added functionalities
