--- profiles/test.yaml
@@ -1,318 +1,0 @@
-check_unique_id:
-  description: 'devansh success test'
-  tag: 'ADOBE-00041'
-
-  # sub_check: true
-
-  # failure_reason: 'fail reason'
-
-  # invert_result: true
-
-  implementations:
-    - filter:
-        grains: 'G@osfinger:CentOS*Linux-8'
-
-      # hubble_version: '>=4.5'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/sshd_config.1
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-
-    - filter:
-        grains: 'G@osfinger:CentOS*Linux-7'
-
-      hubble_version: '>3 AND <7 AND <8'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/ssh_config
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-    - filter:
-        grains: '*'
-
-      # hubble_version: '>=4.5'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/ssh_config
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-    - filter:
-        grains: '*'
-
-      # hubble_version: '>=4.5'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/sshd_config
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-
-          # success_on_file_missing: True
-check_unique_id2:
-  description: 'devansh failure test'
-  tag: 'ADOBE-00042'
-
-  # sub_check: true
-
-  # failure_reason: 'fail reason'
-
-  # invert_result: true
-
-  implementations:
-    - filter:
-        grains: 'G@osfinger:CentOS*Linux-8'
-
-      # hubble_version: '>=4.5'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/sshd_config.1
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-
-    - filter:
-        grains: 'G@osfinger:CentOS*Linux-7'
-
-      hubble_version: '>3 AND <7 AND <8'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/ssh_config1
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-    - filter:
-        grains: '*'
-
-      # hubble_version: '>=4.5'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/ssh_config
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-    - filter:
-        grains: '*'
-
-      # hubble_version: '>=4.5'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/sshd_config
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-
-          # success_on_file_missing: True
-check_unique_id3:
-  description: 'devansh skipped test'
-  tag: 'ADOBE-00043'
-
-  # sub_check: true
-
-  # failure_reason: 'fail reason'
-
-  # invert_result: true
-
-  implementations:
-    - filter:
-        grains: 'G@osfinger:CentOS*Linux-7'
-
-      hubble_version: '>5 AND <7 AND <8'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/ssh_config1
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-    - filter:
-        grains: '*'
-
-      # hubble_version: '>=4.5'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/sshd_config
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-
-          # success_on_file_missing: True
-check_unique_id4:
-  description: 'devansh error test'
-  tag: 'ADOBE-00044'
-
-  # sub_check: true
-
-  # failure_reason: 'fail reason'
-
-  # invert_result: true
-
-  implementations:
-    - filter:
-        grains: 'G@osfinger:CentOS*Linux-7'
-
-      hubble_version: '>3 AND <7 AND <8'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/ssh_config1
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          allow_more_strict: true
-    - filter:
-        grains: '*'
-
-      # hubble_version: '>=4.5'
-
-      # return_no_exec: true
-
-      module: stat
-
-      # type: and
-
-      checks:
-        - path: /etc/ssh/sshd_config
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-
-          # success_on_file_missing: True
-check_unique_id5:
-  description: 'devansh type test'
-  tag: 'ADOBE-00045'
-
-  # sub_check: true
-
-  # failure_reason: 'fail reason'
-
-  # invert_result: true
-
-  implementations:
-    - filter:
-        grains: 'G@osfinger:CentOS*Linux-7'
-
-      hubble_version: '>3 AND <7 AND <8'
-
-      # return_no_exec: true
-
-      module: stat
-
-      type: and
-
-      checks:
-        - path: /etc/ssh/ssh_config
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
-        - path: /etc/ssh/ssh_config
-          gid: 0
-          group: root
-          mode: 644
-          uid: 0
-          user: root
-          allow_more_strict: true
